New feature - upload image, add permission. Truth Twitter v1.1.0

+++ twitter-backend/functions/controller/tweet.js
@@ -1,25 +1,35 @@
 const express = require('express')
+const mongoose = require('mongoose');
 const router = express.Router()
 
 const TweetAccessor = require('../model/tweet.model')
+const UserAccessor = require('../model/user.model')
 
 router.post('/', (req, res) => {
   const data = req.body
-  //   console.log('sign in ' + data)
+  if (!mongoose.Types.ObjectId.isValid(data.userId)) {
+    res.status(404).send('User Id is invalid')
+  }
   // search whether the username and password exit in mongodb
+  UserAccessor.findUserById(data.userId).then( 
+    user => {
+      if(user) {
+        TweetAccessor.insertTweet(data).then(
+          (response) => {
+            res.status(200).send(response)
+          },
+          (error) => res.status(404).send(`Error finding User:${error}`),
+      )
+      } else {
+        res.status(404).send('Cannot find this user')
+      }
+    }
+  ).catch((error) => console.log(error))
 
-  TweetAccessor.insertTweet(data).then(
-      (response) => {
-        res.status(200).send(response)
-      },
-      (error) => res.status(404).send(`Error finding User:${error}`),
-  )
+
 })
 
 router.get('/all', (req, res) => {
-  //   console.log('sign in ' + data)
-  // search whether the username and password exit in mongodb
-
   TweetAccessor.getAllTweet().then(
       (response) => res.status(200).send(response),
       (error) => res.status(404).send(`Error finding User:${error}`),
@@ -27,9 +37,6 @@ router.get('/all', (req, res) => {
 })
 
 router.get('/user/:userId', (req, res) => {
-  //   console.log('sign in ' + data)
-  // search whether the username and password exit in mongodb
-
   TweetAccessor.findTweetByUserId(req.params.userId).then(
       (response) => res.status(200).send(response),
       (error) => res.status(404).send(`Error finding User:${error}`),
@@ -37,9 +44,6 @@ router.get('/user/:userId', (req, res) => {
 })
 
 router.get('/:tweetId', (req, res) => {
-  //   console.log('sign in ' + data)
-  // search whether the username and password exit in mongodb
-
   TweetAccessor.findTweetById(req.params.tweetId).then(
       (response) => res.status(200).send(response),
       (error) => res.status(404).send(`Error finding User:${error}`),
@@ -47,20 +51,44 @@ router.get('/:tweetId', (req, res) => {
 })
 
 router.delete('/:tweetId', (req, res) => {
-  TweetAccessor.deleteTweetById(req.params.tweetId).then(
-      (response) => res.status(200).send(response),
-      (error) => res.status(404).send(`Error finding User:${error}`),
-  )
+  // check if the you are the one who post the tweet
+  if (!mongoose.Types.ObjectId.isValid(req.params.tweetId)) {
+    res.status(404).send('Tweet Id is invalid')
+  }
+  TweetAccessor.findTweetById(req.params.tweetId).then(tweet => {
+    if(tweet) {
+      if(tweet.userId !== req.body.userId) res.status(404).send("Yout don't have the permission to delete other user's post")
+      else {
+        TweetAccessor.deleteTweetById(req.params.tweetId).then(
+          (response) => res.status(200).send(response),
+          (error) => res.status(404).send(`Error finding User:${error}`),
+      )
+      }
+    } else {
+      res.status(404).send('Cannot find this tweet')
+    }
+  })
 })
 
 router.put('/:tweetId', (req, res) => {
-  //   console.log('sign in ' + data)
-  // search whether the username and password exit in mongodb
-
-  TweetAccessor.updateTweetById(req.params.tweetId, req.body).then(
-      (response) => res.status(200).send(response),
-      (error) => res.status(404).send(`Error finding User:${error}`),
-  )
+ // check if the you are the one who post the tweet
+ if (!mongoose.Types.ObjectId.isValid(req.params.tweetId)) {
+  res.status(404).send('Tweet Id is invalid')
+}
+TweetAccessor.findTweetById(req.params.tweetId).then(tweet => {
+  if(tweet) {
+    if(tweet.userId !== req.body.userId) res.status(404).send("Yout don't have the permission to edit other user's post")
+    else {
+      TweetAccessor.updateTweetById(req.params.tweetId, req.body).then(
+        (response) => res.status(200).send(response),
+        (error) => res.status(404).send(`Error finding User:${error}`),
+    )
+    }
+  } else {
+    res.status(404).send('Cannot find this tweet')
+  }
+})
+  
 })
 
 module.exports = router
